pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["Jason Tolbert <jasonalantolbert@gmail.com>"]
description = ""
name = "replitbadger"
version = "0.1.0"
[tool.poetry.dependencies]
Django = "^3.1.3"
django-crispy-forms = "^1.10.0"
django-npm = "^1.0.0"
python = "^3.8"
python-dotenv = "^0.15.0"
[tool.poetry.dev-dependencies]
